refactor: move general components
