+++ src/components/Error/Empty-photos.jsx
@@ -1,4 +1,4 @@
-import styles from '../styles/emptyAndError'
+import styles from './empty-photos'
 
 import React from 'react'
 import { translate } from 'cozy-ui/react/I18n'
@@ -6,8 +6,8 @@ import { translate } from 'cozy-ui/react/I18n'
 export const Empty = ({ t }) => {
   return (
     <div className={styles['pho-empty']}>
-      <h2>{t(`Empty.album_title`)}</h2>
-      <p>{t(`Empty.album_text`)}</p>
+      <h2>{t(`Empty.albums_title`)}</h2>
+      <p>{t(`Empty.albums_text`)}</p>
     </div>
   )
 }
