+++ src/components/Error/ErrorComponent.jsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { translate } from 'cozy-ui/react/I18n'
 import classNames from 'classnames'
 
-import styles from '../styles/emptyAndError'
+import styles from './empty-photos'
 
 export const ErrorComponent = ({ t, errorType }) => {
   return (
