+++ src/components/Error/ErrorShare-drive.jsx
@@ -1,13 +1,13 @@
-import styles from '../styles/emptyAndError'
+import styles from './empty-drive'
 
 import React from 'react'
 import { translate } from 'cozy-ui/react/I18n'
 
 export const ErrorShare = ({ t, errorType }) => {
   return (
-    <div className={styles['pho-errorShare']}>
-      <h2>{t(`Error.${errorType}_title`)}</h2>
-      <p>{t(`Error.${errorType}_text`)}</p>
+    <div className={styles['fil-errorShare']}>
+      <h2>{t(`error.${errorType}_title`)}</h2>
+      <p>{t(`error.${errorType}_text`)}</p>
     </div>
   )
 }
