+++ src/components/Error/empty-drive.styl
@@ -38,12 +38,12 @@ $empty
 .fil-empty
     @extends $empty
     &:before
-        background  embedurl('../assets/icons/icon-type-folder.svg') center center / cover no-repeat
+        background  embedurl('../../drive/assets/icons/icon-type-folder.svg') center center / cover no-repeat
 
 .fil-trash-empty
     @extends $empty
     &:before
-        background  embedurl('../assets/icons/icon-trash-big.svg') center center / cover no-repeat
+        background  embedurl('../../drive/assets/icons/icon-trash-big.svg') center center / cover no-repeat
         height      10rem
 
         @media (max-width: (1023/basefont)rem)
@@ -52,4 +52,4 @@ $empty
 .fil-errorShare
     @extends $empty
     &:before
-        background  embedurl('../assets/icons/icon-cloud-wrong.svg') center center / contain no-repeat
+        background  embedurl('../../drive/assets/icons/icon-cloud-wrong.svg') center center / contain no-repeat
