+++ src/components/Error/empty-photos.styl
@@ -42,14 +42,14 @@ $empty
 .pho-error
     @extends $empty
     &:before
-        background  embedurl('../assets/icons/icon-image-broken.svg') center center / cover no-repeat
+        background  embedurl('../../photos/assets/icons/icon-image-broken.svg') center center / cover no-repeat
 
 .pho-errorShare
     @extends $empty
     &:before
-        background  embedurl('../assets/icons/icon-cloud-wrong.svg') center center / contain no-repeat
+        background  embedurl('../../photos/assets/icons/icon-cloud-wrong.svg') center center / contain no-repeat
 
 .pho-empty
     @extends $empty
     &:before
-        background  embedurl('../assets/icons/icon-main-app.svg') center center / cover no-repeat
+        background  embedurl('../../photos/assets/icons/icon-main-app.svg') center center / cover no-repeat
