+++ src/components/Error/oops.styl
@@ -1,0 +1,8 @@
+@require 'cozy-ui'
+@require './empty-drive'
+
+.fil-oops
+    @extends $empty
+
+    &:before
+        background  embedurl("../../drive/assets/icons/icon-folder-broken.svg") center center / cover no-repeat
