+++ src/drive/components/LightFolderView.jsx
@@ -7,7 +7,7 @@ import FolderContent from './FolderContent'
 import FileListHeader from './FileListHeader'
 import Topbar from './Topbar'
 import Breadcrumb from '../containers/Breadcrumb'
-import ErrorShare from './ErrorShare'
+import ErrorShare from 'components/Error/ErrorShare-drive'
 
 import DownloadButton from './DownloadButton'
 import Menu, { Item } from 'components/Menu'
