--- src/drive/styles/oops.styl
@@ -1,8 +1,0 @@
-@require 'cozy-ui'
-@require './empty'
-
-.fil-oops
-    @extends $empty
-
-    &:before
-        background  embedurl("../assets/icons/icon-folder-broken.svg") center center / cover no-repeat
